+++ R/cnt_skeleton.R
@@ -1,17 +1,20 @@
 #' Create a skeleton of a closed polygon object
 #'
 #' @param input \code{sf}, \code{sfc} or \code{SpatVector} polygons object
-#' @param keep numeric, proportion of points to retain (0.05-Inf; default 1).
+#' @param keep numeric, proportion of points to retain (0.05-Inf; default 0.1).
 #' See Details.
 #'
 #' @details
 #' - If \code{keep} equals 1 (default), no transformation will occur. The
 #' function will use the original geometry to find the skeleton.
 #'
-#' - If the keep parameter is below 1, then the [rmapshaper::ms_simplify()]
-#' function will be used with Douglas-Peuker algorithm. So the original input
+#' - If the \code{keep} parameter is below 1, then the [geos::geos_simplify()]
+#' function will be used. So the original input
 #' geometry would be simplified, and the resulting skeleton will be cleaner but
 #' maybe more edgy.
+#' The current realisation of simplification is similar (but not identical) to
+#' [rmapshaper::ms_simplify()] one with Douglas-Peuker algorithm. However,
+#' due to \code{geos} superpower, 50x times faster.
 #'
 #' - If the \code{keep} is above 1, then the densification
 #' algorithm is applied using the [geos::geos_densify()] function. This may
@@ -39,14 +42,24 @@
 #' plot(pol_skeleton)
 cnt_skeleton <-
   function(input,
-           keep = 0.1) {
+           keep = 0.5) {
     UseMethod("cnt_skeleton")
+  }
+
+#' @export
+cnt_skeleton.geos_geometry <-
+  function(input,
+           keep = 0.5) {
+    cnt_skeleton_geos(
+      input = input,
+      keep = keep
+    )
   }
 
 #' @export
 cnt_skeleton.sf <-
   function(input,
-           keep = 0.1) {
+           keep = 0.5) {
     cnt_skeleton_sf(
       input = input,
       keep = keep
@@ -56,7 +69,7 @@ cnt_skeleton.sf <-
 #' @export
 cnt_skeleton.sfc <-
   function(input,
-           keep = 0.1) {
+           keep = 0.5) {
     cnt_skeleton_sfc(
       input = input,
       keep = keep
@@ -66,77 +79,32 @@ cnt_skeleton.sfc <-
 #' @export
 cnt_skeleton.SpatVector <-
   function(input,
-           keep = 0.1) {
+           keep = 0.5) {
     cnt_skeleton_terra(
       input = input,
       keep = keep
     )
   }
 
-#' @export
-cnt_skeleton.character <-
+cnt_skeleton_geos <-
   function(input,
-           keep = 0.1) {
-    cnt_skeleton_terra(
-      input = input,
-      keep = keep
-    )
-  }
-
-
-cnt_skeleton_sfc <-
-  function(input,
-           keep = 0.1) {
-    # Check if input is of class 'sf' or 'sfc' and 'POLYGON'
-    stopifnot(check_sf_polygon(input))
-
-    # Save CRS
-    crs <-
-      sf::st_crs(input)
-
-    # Transform to GEOS geometry
-    input_geos <-
-      input |>
-      geos::as_geos_geometry()
-
+           keep = 0.5) {
     # Simplify or densify or do nothing
     if (keep == 1) {
-      pol_geos <- input_geos
+      pol_geos <- input
     } else if (keep < 1 && keep >= 0.05) {
+      # Simplify
       pol_geos <-
-        rmapshaper::ms_simplify(
-          input,
-          keep = keep,
-          method = "dp",
-          keep_shapes = TRUE
-        ) |>
-        geos::as_geos_geometry()
+        geos_ms_simplify(input, keep = keep)
     } else if (keep < 0.05) {
+      # Simplify at lower bound, otherwise result
+      # might be poor
       pol_geos <-
-        rmapshaper::ms_simplify(
-          input,
-          keep = 0.05,
-          method = "dp",
-          keep_shapes = TRUE
-        ) |>
-        geos::as_geos_geometry()
+        geos_ms_simplify(input, keep = 0.05)
     } else {
-      perimeter_length <-
-        geos::geos_length(input_geos)
-
-      point_count <-
-        input_geos |>
-        geos::geos_unique_points() |>
-        geos::geos_num_coordinates()
-
-      point_density <-
-        perimeter_length / point_count
-
+      # Densify
       pol_geos <-
-        geos::geos_densify(
-          input_geos,
-          tolerance = point_density / keep
-        )
+        geos_ms_densify(input, keep = keep)
     }
 
     # Find polygon skeleton
@@ -146,17 +114,37 @@ cnt_skeleton_sfc <-
       geos::geos_intersection(pol_geos) |>
       geos::geos_unnest(keep_multi = FALSE)
 
+    pol_skeleton
+  }
+
+cnt_skeleton_sfc <-
+  function(input,
+           keep = 0.5) {
+    # Check if input is of class 'sf' or 'sfc' and 'POLYGON'
+    stopifnot(check_sf_polygon(input))
+
+    # Save CRS
+    crs <-
+      sf::st_crs(input)
+
+    # Transform to GEOS geometry
+    input_geos <-
+      geos::as_geos_geometry(input)
+
+    # Find GEOS skeleton
+    pol_skeleton <-
+      cnt_skeleton_geos(input = input_geos, keep = keep)
+
     # Transform back to sf
     pol_skeleton_crs <-
-      pol_skeleton |>
-      sf::st_as_sfc(crs = crs)
+      sf::st_as_sfc(x = pol_skeleton, crs = crs)
 
-    return(pol_skeleton_crs)
+    pol_skeleton_crs
   }
 
 cnt_skeleton_sf <-
   function(input,
-           keep = 0.1) {
+           keep = 0.5) {
     # Check if input is of class 'sf' or 'sfc' and 'POLYGON'
     stopifnot(check_sf_polygon(input))
 
@@ -166,135 +154,36 @@ cnt_skeleton_sf <-
 
     # Transform to GEOS geometry
     input_geos <-
-      input |>
-      geos::as_geos_geometry()
+      geos::as_geos_geometry(input)
 
-    # Simplify or densify or do nothing
-    if (keep == 1) {
-      pol_geos <- input_geos
-    } else if (keep < 1 && keep >= 0.05) {
-      pol_geos <-
-        rmapshaper::ms_simplify(
-          input,
-          keep = keep,
-          method = "dp",
-          keep_shapes = TRUE
-        ) |>
-        geos::as_geos_geometry()
-    } else if (keep < 0.05) {
-      pol_geos <-
-        rmapshaper::ms_simplify(
-          input,
-          keep = 0.05,
-          method = "dp",
-          keep_shapes = TRUE
-        ) |>
-        geos::as_geos_geometry()
-    } else {
-      perimeter_length <-
-        geos::geos_length(input_geos)
-
-      point_count <-
-        input_geos |>
-        geos::geos_unique_points() |>
-        geos::geos_num_coordinates()
-
-      point_density <-
-        perimeter_length / point_count
-
-      pol_geos <-
-        geos::geos_densify(
-          input_geos,
-          tolerance = point_density / keep
-        )
-    }
-
-    # Find polygon skeleton
+    # Find GEOS skeleton
     pol_skeleton <-
-      pol_geos |>
-      geos::geos_voronoi_edges() |>
-      geos::geos_intersection(pol_geos) |>
-      geos::geos_unnest(keep_multi = FALSE)
+      cnt_skeleton_geos(input = input_geos, keep = keep)
 
     # Transform back to sf
     pol_skeleton_crs <-
-      pol_skeleton |>
-      sf::st_as_sf(crs = crs) ## |>
-    # cbind(sf::st_drop_geometry(input))
+      sf::st_as_sf(x = pol_skeleton, crs = crs)
 
-    return(pol_skeleton_crs)
+    pol_skeleton_crs
   }
 
 cnt_skeleton_terra <-
   function(input,
-           keep = 0.1) {
+           keep = 0.5) {
     # Check if input is of class 'SpatVector' and 'polygons'
     stopifnot(check_terra_polygon(input))
 
-    # Save CRS
-    crs <-
-      terra::crs(input)
-
     # Transform to GEOS geometry
     input_geos <-
-      input |>
-      terra::geom(wkt = TRUE) |>
-      geos::as_geos_geometry()
+      terra_to_geos(input)
 
-    # Simplify or densify or do nothing
-    if (keep == 1) {
-      pol_geos <- input_geos
-    } else if (keep < 1 && keep >= 0.05) {
-      pol_geos <-
-        rmapshaper::ms_simplify(
-          sf::st_as_sf(input_geos),
-          keep = keep,
-          method = "dp",
-          keep_shapes = TRUE
-        ) |>
-        geos::as_geos_geometry()
-    } else if (keep < 0.05) {
-      pol_geos <-
-        rmapshaper::ms_simplify(
-          sf::st_as_sf(input_geos),
-          keep = 0.05,
-          method = "dp",
-          keep_shapes = TRUE
-        ) |>
-        geos::as_geos_geometry()
-    } else {
-      perimeter_length <-
-        geos::geos_length(input_geos)
-
-      point_count <-
-        input_geos |>
-        geos::geos_unique_points() |>
-        geos::geos_num_coordinates()
-
-      point_density <-
-        perimeter_length / point_count
-
-      pol_geos <-
-        geos::geos_densify(
-          input_geos,
-          tolerance = point_density / keep
-        )
-    }
-
-    # Find skeleton
+    # Find GEOS skeleton
     pol_skeleton <-
-      pol_geos |>
-      geos::geos_voronoi_edges() |>
-      geos::geos_intersection(pol_geos) |>
-      geos::geos_unnest(keep_multi = FALSE)
+      cnt_skeleton_geos(input = input_geos, keep = keep)
 
     # Transform back to SpatVect
     pol_skeleton_crs <-
-      pol_skeleton |>
-      wk::as_wkt() |>
-      as.character() |>
-      terra::vect(crs = crs)
-    # cbind(terra::as.data.frame(input))
+      geos_to_terra(pol_skeleton)
 
-    return(pol_skeleton_crs)
+    pol_skeleton_crs
   }